+++ src/main/java/com/telepathicgrunt/repurposedstructures/configs/RSOutpostsConfig.java
@@ -26,7 +26,7 @@ public class RSOutpostsConfig implements ConfigData {
         @Comment("How rare are Nether Brick Outposts in non-warped Nether biomes."
                 + "\n1 for spawning in most chunks and 1001 for none.")
         @ConfigEntry.BoundedDiscrete(min = 1, max = 1001)
-        public int netherBrickOutpostSpawnrate = 33;
+        public int netherBrickOutpostSpawnrate = 34;
 
         @ConfigEntry.Gui.Tooltip(count = 2)
         @Comment("Add Nether Brick Outposts to modded Nether biomes that other nether outposts don't fit in.")
@@ -37,7 +37,7 @@ public class RSOutpostsConfig implements ConfigData {
         @Comment("How rare are Warped Outposts in Warped Nether biomes."
                 + "\n1 for spawning in most chunks and 1001 for none.")
         @ConfigEntry.BoundedDiscrete(min = 1, max = 1001)
-        public int warpedOutpostSpawnrate = 33;
+        public int warpedOutpostSpawnrate = 34;
 
         @ConfigEntry.Gui.Tooltip()
         @Comment("Add Warped Outposts to modded Nether Warped biomes.")
@@ -48,7 +48,7 @@ public class RSOutpostsConfig implements ConfigData {
         @Comment("How rare are Crimson Outposts in Warped Nether biomes."
                 + "\n1 for spawning in most chunks and 1001 for none.")
         @ConfigEntry.BoundedDiscrete(min = 1, max = 1001)
-        public int crimsonOutpostSpawnrate = 33;
+        public int crimsonOutpostSpawnrate = 34;
 
         @ConfigEntry.Gui.Tooltip()
         @Comment("Add Crimson Outposts to modded Nether Warped biomes.")