+++ src/main/java/com/telepathicgrunt/repurposedstructures/configs/RSTemplesConfig.java
@@ -28,7 +28,7 @@ public class RSTemplesConfig implements ConfigData {
         @Comment("How rare are Nether Temples in Nether Wastelands."
                 + "\n1 for spawning in most chunks and 1001 for none.")
         @ConfigEntry.BoundedDiscrete(min = 1, max = 1001)
-        public int netherWastelandTempleSpawnrate = 29;
+        public int netherWastelandTempleSpawnrate = 27;
 
         @ConfigEntry.Gui.Tooltip(count = 2)
         @Comment("Add Nether Wasteland Temples to modded Nether biomes that other nether temples don't fit in.")
@@ -39,7 +39,7 @@ public class RSTemplesConfig implements ConfigData {
         @Comment("How rare are Nether Basalt Temples in Nether Basalt Delta biomes."
                 + "\n1 for spawning in most chunks and 1001 for none.")
         @ConfigEntry.BoundedDiscrete(min = 1, max = 1001)
-        public int netherBasaltTempleSpawnrate = 29;
+        public int netherBasaltTempleSpawnrate = 27;
 
         @ConfigEntry.Gui.Tooltip
         @Comment("Add Nether Basalt Temples to modded Nether Basalt biomes.")
@@ -50,7 +50,7 @@ public class RSTemplesConfig implements ConfigData {
         @Comment("How rare are Nether Crimson Temples in Nether Crimson Forest."
                 + "\n1 for spawning in most chunks and 1001 for none.")
         @ConfigEntry.BoundedDiscrete(min = 1, max = 1001)
-        public int netherCrimsonTempleSpawnrate = 29;
+        public int netherCrimsonTempleSpawnrate = 27;
 
         @ConfigEntry.Gui.Tooltip
         @Comment("Add Nether Crimson Temples to modded Nether Crimson Forest biomes.")
@@ -61,7 +61,7 @@ public class RSTemplesConfig implements ConfigData {
         @Comment("How rare are Nether Crimson Temples in Nether Warped Forest."
                 + "\n1 for spawning in most chunks and 1001 for none.")
         @ConfigEntry.BoundedDiscrete(min = 1, max = 1001)
-        public int netherWarpedTempleSpawnrate = 29;
+        public int netherWarpedTempleSpawnrate = 27;
 
         @ConfigEntry.Gui.Tooltip
         @Comment("Add Nether Warped Temples to modded Nether Warped Forest biomes.")
@@ -72,7 +72,7 @@ public class RSTemplesConfig implements ConfigData {
         @Comment("How rare are Nether Soul Temples in Nether Soul Sand Valley."
                 + "\n1 for spawning in most chunks and 1001 for none.")
         @ConfigEntry.BoundedDiscrete(min = 1, max = 1001)
-        public int netherSoulTempleSpawnrate = 29;
+        public int netherSoulTempleSpawnrate = 27;
 
         @ConfigEntry.Gui.Tooltip
         @Comment("Add Nether Soul Temples to modded Nether Soul Sand Valley biomes.")