+++ src/main/java/com/telepathicgrunt/repurposedstructures/configs/RSVillagesConfig.java
@@ -67,11 +67,11 @@ public class RSVillagesConfig implements ConfigData {
     @Comment("How rare are Crimson Village in Crimson Forest biomes."
             + "\n1 for spawning in most chunks and 1001 for none.")
     @ConfigEntry.BoundedDiscrete(min = 1, max = 1001)
-    public int crimsonVillageSpawnrate = 27;
+    public int crimsonVillageSpawnrate = 30;
 
     @ConfigEntry.Gui.Tooltip(count = 2)
     @Comment("How rare are Warped Village in Warped Forest biomes."
             + "\n1 for spawning in most chunks and 1001 for none.")
     @ConfigEntry.BoundedDiscrete(min = 1, max = 1001)
-    public int warpedVillageSpawnrate = 27;
+    public int warpedVillageSpawnrate = 30;
 }